+++ programs/on-chain-twitter/src/lib.rs
@@ -18,7 +18,7 @@ pub mod on_chain_twitter {
     pub fn transfer_ownership_user_account(ctx: Context<TransferOwnershipUserAccount>, new_owner: Pubkey) -> Result<()>{
         instructions::transfer_ownership_user_account::handler(ctx, new_owner)
     }
-    pub fn get_number_of_tweets_by_user(ctx: Context<GetNumberOfTweetsByUser>) -> Result<u64>{
+    pub fn get_number_of_tweets_by_user(ctx: Context<GetNumberOfTweetsByUser>) -> Result<u8>{
         instructions::get_number_of_tweets_by_user::handler(ctx)
     }
     pub fn delete_twitter_account(_ctx: Context<DeleteTwitterAccount>) -> Result<()> {
